Create ClientSideIterator for BatchScanner. Use for gathering entries at client.
ClientSideIteratorAggregatingScanner reads all the entries from a scan into memory at the client, then runs them through iterators at the client. A use case is the "plus op" performed on a remote table, now performed at the client when we don't want to write results to another table.

Found a bug in SeekFilterIterator. Abstracted HardListIterator to MapIterator, which converts a SortedMap into an SKVI.