Make iterator priority and name first class in DynamicIteratorSetting.
Now maintained even when they don't have to be, because it is safer
to over-specify than under. Next step is to expose
the iterator name to the Graphulo One/TwoTable API.
But what if users want to replace an existing iterator?
Maybe allow specifying name, and make usrs remove iterators they want to replace.