DynamicIterator capability to run at certain scopes. Fix kTrussAdj.
kTrussAdj has the chance of being incorrect for k>3 and
large enough graphs that there are flushes in the middle of the first multiply.
This fixes that situation to only apply the filter after all entries
are written, and only when all entries are visible (scan and full majc).

DynamicIterator can now control what scopes its component iterators apply.