Fix RemoteWriteIterator bug. Includes print debug stmts.
This bug causes the kTrussAdj algorithm to run duplicate computation.
It is due to the SourceSwitchingIterator in Accumulo
tearing down the Graphulo iterator stack after it has emitted
its one and only one entry (with monitoring turned off).
This is now fixed by placing the number of ranges that have been completed
inside the column visibility of monitoring entries
emitted by RemoteWriteIterator.